RAG_project/requirements.txt
langchain
sentence-transformers
faiss-cpu
huggingface_hub
transformers
torch
tqdm
numpy

pip install -r requirements.txt